+++ gluecodium/src/test/resources/smoke/comments/output/swift/smoke/DeprecationCommentsOnly.swift
@@ -83,6 +83,10 @@ public enum SomeEnum : UInt32, CaseIterable, Codable {
 public struct SomeStruct {
     @available(*, deprecated, message: "Unfortunately, this field is deprecated.")
     public var someField: DeprecationCommentsOnly.Usefulness
+    public init() {
+        self.someField = false
+    }
+    @available(*, deprecated)
     public init(someField: DeprecationCommentsOnly.Usefulness = false) {
         self.someField = someField
     }